Fix warnings #5663

+++ src/condor_utils/file_transfer.cpp
@@ -2429,7 +2429,7 @@ FileTransfer::DoDownload( filesize_t *total_bytes, ReliSock *s)
 		jobAd.LookupInteger(ATTR_PROC_ID, proc);
 
 		dprintf(D_STATS, "File Transfer Download: JobId: %d.%d bytes: %lld seconds: %.1f dest: %s %s\n", 
-			cluster, proc, *total_bytes, (double)(downloadEndTime - downloadStartTime), s->peer_ip_str(), (stats ? stats : "") );
+			cluster, proc, (long long)*total_bytes, (double)(downloadEndTime - downloadStartTime), s->peer_ip_str(), (stats ? stats : "") );
 	}
 
 
@@ -3817,7 +3817,7 @@ FileTransfer::ExitDoUpload(filesize_t *total_bytes, ReliSock *s, priv_state save
 
 		char *stats = s->get_statistics();
 		dprintf(D_STATS, "File Transfer Upload: JobId: %d.%d bytes: %lld seconds: %.1f dest: %s %s\n", 
-			cluster, proc, *total_bytes, (uploadEndTime - uploadStartTime), s->peer_ip_str(), (stats ? stats : "") );
+			cluster, proc, (long long)*total_bytes, (double) (uploadEndTime - uploadStartTime), s->peer_ip_str(), (stats ? stats : "") );
 	}
 
 	return rc;
